+++ ImageLounge/src/DkGui/DkWidgets.h
@@ -449,12 +449,12 @@ public:
 	bool isAutoHide() const;
 
 signals:
-	void zoomSignal(float zoomLevel);
+	void zoomSignal(double zoomLevel);
 
 public slots:
 	virtual void setVisible(bool visible, bool autoHide = false);
 
-	void updateZoom(float zoomLevel);
+	void updateZoom(double zoomLevel);
 	void on_sbZoom_valueChanged(double zoomLevel);
 	void on_slZoom_valueChanged(int zoomLevel);
